certificate/requirement.txt
weasyprint
yagmail
mysql-connector-python
wand
pillow